v3

+++ Cooking bot/Models/Recipe.cs
@@ -12,6 +12,6 @@ namespace CookingBot.Models
         public string Title { get; set; } = string.Empty;
         public List<string> Ingredients { get; set; } = new();
         public string Instructions { get; set; } = string.Empty;
-        public int Calories { get; set; } = 0;
+        public double Calories { get; set; } = 0;
     }
 }
